+++ src/AppBundle/Repository/SaleDeliveryNoteRepository.php
@@ -2,7 +2,10 @@
 
 namespace AppBundle\Repository;
 
+use AppBundle\Entity\Enterprise;
 use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\Query;
+use Doctrine\ORM\QueryBuilder;
 
 /**
  * Class SaleDeliveryNoteRepository.
@@ -11,4 +14,64 @@ use Doctrine\ORM\EntityRepository;
  */
 class SaleDeliveryNoteRepository extends EntityRepository
 {
+    /**
+     * @return QueryBuilder
+     */
+    public function getEnabledSortedByNameQB()
+    {
+        return $this->createQueryBuilder('s')
+            ->where('s.enabled = :enabled')
+            ->setParameter('enabled', true)
+            ->orderBy('s.date', 'DESC')
+        ;
+    }
+
+    /**
+     * @return Query
+     */
+    public function getEnabledSortedByNameQ()
+    {
+        return $this->getEnabledSortedByNameQB()->getQuery();
+    }
+
+    /**
+     * @return array
+     */
+    public function getEnabledSortedByName()
+    {
+        return $this->getEnabledSortedByNameQ()->getResult();
+    }
+
+    /**
+     * @param Enterprise $enterprise
+     *
+     * @return QueryBuilder
+     */
+    public function getFilteredByEnterpriseSortedByNameQB(Enterprise $enterprise)
+    {
+        return $this->getEnabledSortedByNameQB()
+            ->andWhere('s.enterprise = :enterprise')
+            ->setParameter('enterprise', $enterprise)
+        ;
+    }
+
+    /**
+     * @param Enterprise $enterprise
+     *
+     * @return Query
+     */
+    public function getFilteredByEnterpriseSortedByNameQ(Enterprise $enterprise)
+    {
+        return $this->getFilteredByEnterpriseSortedByNameQB($enterprise)->getQuery();
+    }
+
+    /**
+     * @param Enterprise $enterprise
+     *
+     * @return array
+     */
+    public function getFilteredByEnterpriseSortedByName(Enterprise $enterprise)
+    {
+        return $this->getFilteredByEnterpriseSortedByNameQ($enterprise)->getResult();
+    }
 }